backend/requirements.txt
scikit-learn>=1.3,<2.0
matplotlib>=3.8,<4.0
fastapi>=0.111,<1.0
uvicorn[standard]>=0.29,<1.0
PyNaCl>=1.5,<2.0
slowapi>=0.1.9,<1.0
base58>=2.1,<3.0
